cnpj/requirements.txt
# Standalone CNPJ module dependencies
sqlalchemy>=2.0
psycopg2-binary
python-dotenv
httpx
